fix: clippy

+++ bottlecap/src/metrics/enhanced/lambda.rs
@@ -863,7 +863,7 @@ mod tests {
         lambda.set_runtime_done_metrics(
             &RuntimeDoneMetrics {
                 duration_ms: 100.0,
-                produced_bytes: Some(42 as u64),
+                produced_bytes: Some(42_u64),
             },
             now,
         );
@@ -982,7 +982,7 @@ mod tests {
         let lambda = Lambda::new(metrics_aggr.clone(), my_config);
         let runtime_done_metrics = RuntimeDoneMetrics {
             duration_ms: 100.0,
-            produced_bytes: Some(42 as u64),
+            produced_bytes: Some(42_u64),
         };
         let now: i64 = std::time::UNIX_EPOCH
             .elapsed()
